src/requirements.txt
boto3==1.26.45
pandas==1.5.2
psycopg2-binary==2.9.5